pyproject.toml
[project]
name = "app-2-flatmates-bill"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fpdf>=1.7.2",
    "python-dotenv>=1.0.1",
    "pyuploadcare>=6.0.0",
    "space-tracer>=4.12.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]
